% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_uncertainty.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Compute the cosine similarity between two sentences}
\usage{
cosine_similarity(sentence1, sentence2)
}
\arguments{
\item{sentence1}{the first sentence}

\item{sentence2}{the second sentence}
}
\description{
Compute the cosine similarity between two sentences
}
