% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trustme_models.R
\name{generate_embeddings}
\alias{generate_embeddings}
\title{Generate embeddings from a list of texts}
\usage{
generate_embeddings(text_list, model = "bert-base-uncased")
}
\arguments{
\item{text_list}{a list of texts}

\item{model}{Character string specifying pre-trained language model (default 'bert-base-uncased'). For full list of options see pretrained models at \href{https://huggingface.co/}{HuggingFace}. For example use "bert-base-multilingual-cased", "openai-gpt", "gpt2", "ctrl", "transfo-xl-wt103", "xlnet-base-cased", "xlm-mlm-enfr-1024", "distilbert-base-cased", "roberta-base", or "xlm-roberta-base". Only load models that you trust from HuggingFace; loading a malicious model can execute arbitrary code on your computer).}
}
\value{
a dataframe containing the embedding of all texts submitted to the function
}
\description{
Generate embeddings from a list of texts
}
