% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_uncertainty.R
\name{plot_uncertainty}
\alias{plot_uncertainty}
\title{Plot uncertainty of LLM responses}
\usage{
plot_uncertainty(res_uncertainty)
}
\arguments{
\item{res_uncertainty}{The result of compute_ncertainty function}
}
\value{
a list containing
\itemize{
\item plot_graph: a plot of the graph with similarity between responses as adjacency matrix
\item plot_eccentricity: a barplot displaying the eccentrictiy of each response
\item plot_degree: a barplot displaying the degree (\eqn{\approx} confidence) in each response
\item plot_uni_metrics: a barplot displaying different univariate metrics
}
}
\description{
Plot uncertainty of LLM responses
}
\examples{
\dontrun{
# Processing time is often longer than ten seconds
# because the function uses multiple times a large language model.
library(NaileR)
data(beard_cont)
intro_beard <- 'A survey was conducted about beards
and 8 types of beards were described.
In the data that follow, beards are named B1 to B8.'
intro_beard <- gsub('\n', ' ', intro_beard) |>
  stringr::str_squish()

req_beard <- 'Please give a name to each beard
and summarize what makes this beard unique.'
req_beard <- gsub('\n', ' ', req_beard) |>
  stringr::str_squish()

res_trustme <- trustme_models(res_beard, models = c("llama2", "llama3", "llama3.1","mistral"))

uncertainty <- compute_uncertainty_metrics(unlist(res_trustme$responses))
plot_uq <- plot_uncertainty(uncertainty)


}

}
