% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trustme_models.R
\name{trustme_models}
\alias{trustme_models}
\title{Evaluate uncertainty of LLM responses (one model and multiple repetitions, or multiple models and on repetition)}
\usage{
trustme_models(prompt, models = NULL, num_repeats = 1)
}
\arguments{
\item{prompt}{a prompt to be given to the LLM}

\item{models}{A vector of characters strings containing names of different LLM that have to be tested. If length(models) > 1, then num_repeats has to be equal to 1 (either multiple models with one repetition each or one model with multiple repetitions.)}

\item{num_repeats}{an integer specifying the number of times you want to evaluate the same prompt. If num_repeats > 1, then length(models) has to be equal to 1 (either multiple models with one repetition each or one model with multiple repetitions.)}
}
\value{
a list containing
\itemize{
\item responses: a named list of all responses of all models or repetitions
\item central_answer: an answer meant to be the most representative one across all models or repetitions
\item pca: the result of a PCA computed on the embeddings of all responses
\item plot_pca: the two first principal components of the PCA
}
}
\description{
Evaluate uncertainty of LLM responses (one model and multiple repetitions, or multiple models and on repetition)
}
\examples{
\dontrun{
# Processing time is often longer than ten seconds
# because the function uses multiple times a large language model.

data(beard_cont)
intro_beard <- 'A survey was conducted about beards
and 8 types of beards were described.
In the data that follow, beards are named B1 to B8.'
intro_beard <- gsub('\n', ' ', intro_beard) |>
  stringr::str_squish()

req_beard <- 'Please give a name to each beard
and summarize what makes this beard unique.'
req_beard <- gsub('\n', ' ', req_beard) |>
  stringr::str_squish()

res_trustme <- trustme_models(res_beard, models = c("llama2", "llama3", "llama3.1","mistral"))
names(res_trustme)
res_trustme$central_answer


}

}
