% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_uncertainty.R
\name{compute_uncertainty_metrics}
\alias{compute_uncertainty_metrics}
\title{Measure uncertainty of LLM responses}
\usage{
compute_uncertainty_metrics(vec_text)
}
\arguments{
\item{vec_text}{A vector of text for which the global uncertainty has to be measured}
}
\value{
a list containing
\itemize{
\item laplacian_normalized: the normalized Laplacian of the similarity matrix
\item similarity_matrix: the similarity matrix
\item W: the symetrized similarity matrix
\item UeigV: The "number of semantic meanings"
\item eigen_values: the eigen values of the Laplacian
\item Udeg: The average pairwise distance between responses
\item Cdeg: for each node, the confidence score
\item Uecc: average eccentricity
\item Cecc: eccentricity for each node (useful for comparison between nodes)
}
}
\description{
Measure uncertainty of LLM responses
}
\details{
This function computes metrics based on the article of Lin et al., 2024. It embeds responses using the
\href{https://huggingface.co/sentence-transformers/all-mpnet-base-v2}{all-mpnet-base-v2} model, and then compute cosine similarities between them.
Several metrics are derived from the adjacency matrix built from the cosine similarities. Some are designed to quantify the global uncertainty (UeigV, Udeg, Uecc)
and other are designed to quantify and compare the uncertainty of each node (Cdeg, Cecc).
}
\examples{
\dontrun{
# Processing time is often longer than ten seconds
# because the function uses multiple times a large language model.
library(NaileR)
data(beard_cont)
intro_beard <- 'A survey was conducted about beards
and 8 types of beards were described.
In the data that follow, beards are named B1 to B8.'
intro_beard <- gsub('\n', ' ', intro_beard) |>
  stringr::str_squish()

req_beard <- 'Please give a name to each beard
and summarize what makes this beard unique.'
req_beard <- gsub('\n', ' ', req_beard) |>
  stringr::str_squish()

res_trustme <- trustme_models(res_beard, models = c("llama2", "llama3", "llama3.1","mistral"))

uncertainty <- compute_uncertainty_metrics(unlist(res_trustme$responses))


}

}
\references{
Lin, Z., Trivedi, S., & Sun, J. (2024). Generating with confidence: Uncertainty quantification for black-box large language models
}
