% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrustMe-package.R
\docType{package}
\name{TrustMe-package}
\alias{TrustMe}
\alias{TrustMe-package}
\title{TrustMe: TrustMe: Turning to R for Understanding Semantic Topology from Model’s Embeddings}
\description{
An R package designed to enhance confidence in LLM-generated responses, particularly within the NaileR ecosystem. TrustMe offers different strategies to evaluate the reliability of LLM-generated responses (cross-model, internal-model and data-model consistency).
}
\author{
\strong{Maintainer}: Sébastien Lê \email{sebastien.le@institut-agro.fr} (\href{https://orcid.org/0000-0001-8814-6714}{ORCID})

Authors:
\itemize{
  \item Rémi Mahmoud (\href{https://orcid.org/0000-0003-0853-0834}{ORCID})
}

}
\keyword{internal}
